statsPipeline/requirements.txt
pandas>=1.5.0
mysql-connector-python>=8.0.0
requests>=2.28.0